nf[Gestures]: extract tap gestures
From GestureProcessor -> GestureMapper

Related: 6052
