+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/GestureProcessor.java
@@ -20,51 +20,16 @@ import android.content.SharedPreferences;
 
 import com.ichi2.anki.reviewer.GestureMapper;
 
-import static com.ichi2.anki.cardviewer.ViewerCommand.COMMAND_NOTHING;
-
 public class GestureProcessor {
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapLeft;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapRight;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapTop;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapBottom;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapTopLeft;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapTopRight;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapCenter;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapBottomLeft;
-    @ViewerCommand.ViewerCommandDef
-    private int mGestureTapBottomRight;
     @ViewerCommand.ViewerCommandDef
     private int mGestureDoubleTap;
     @ViewerCommand.ViewerCommandDef
     private int mGestureLongclick;
 
-    private boolean mUseCornerTouch;
-
-    private GestureMapper mGestureMapper = new GestureMapper();
+    private final GestureMapper mGestureMapper = new GestureMapper();
 
 
     public void init(SharedPreferences preferences) {
-        mGestureTapLeft = Integer.parseInt(preferences.getString("gestureTapLeft", "3"));
-        mGestureTapRight = Integer.parseInt(preferences.getString("gestureTapRight", "6"));
-        mGestureTapTop = Integer.parseInt(preferences.getString("gestureTapTop", "12"));
-        mGestureTapBottom = Integer.parseInt(preferences.getString("gestureTapBottom", "2"));
-
-        mUseCornerTouch = preferences.getBoolean("gestureCornerTouch", false);
-        if (mUseCornerTouch) {
-            mGestureTapTopLeft = Integer.parseInt(preferences.getString("gestureTapTopLeft", "0"));
-            mGestureTapTopRight = Integer.parseInt(preferences.getString("gestureTapTopRight", "0"));
-            mGestureTapCenter = Integer.parseInt(preferences.getString("gestureTapCenter", "0"));
-            mGestureTapBottomLeft = Integer.parseInt(preferences.getString("gestureTapBottomLeft", "0"));
-            mGestureTapBottomRight = Integer.parseInt(preferences.getString("gestureTapBottomRight", "0"));
-        }
         mGestureDoubleTap = Integer.parseInt(preferences.getString("gestureDoubleTap", "7"));
         mGestureLongclick = Integer.parseInt(preferences.getString("gestureLongclick", "11"));
         mGestureMapper.init(preferences);
@@ -72,53 +37,8 @@ public class GestureProcessor {
 
     @ViewerCommand.ViewerCommandDef
     public int getCommandFromTap(int height, int width, float posX, float posY) {
-        if (width == 0 || height == 0) {
-            return COMMAND_NOTHING;
-        }
-
-        if (mUseCornerTouch) {
-            return getCornerTouchCommand(height, width, posX, posY);
-        }
-        return getFourCornerTap(height, width, posX, posY);
+        return mGestureMapper.getCommandFromTap(height, width, posX, posY);
     }
-
-
-    public int getCornerTouchCommand(int height, int width, float posX, float posY) {
-        GestureSegment segment = GestureSegment.fromTap(height, width, posX, posY);
-
-        switch (segment) {
-            case TOP_LEFT: return mGestureTapTopLeft;
-            case TOP_CENTER: return mGestureTapTop;
-            case TOP_RIGHT: return mGestureTapTopRight;
-            case MIDDLE_LEFT: return mGestureTapLeft;
-            case MIDDLE_CENTER: return mGestureTapCenter;
-            case MIDDLE_RIGHT: return mGestureTapRight;
-            case BOTTOM_LEFT: return mGestureTapBottomLeft;
-            case BOTTOM_CENTER: return mGestureTapBottom;
-            case BOTTOM_RIGHT: return mGestureTapBottomRight;
-            default: throw new IllegalStateException("invalid switch");
-        }
-
-    }
-
-
-    private int getFourCornerTap(int height, int width, float posX, float posY) {
-        boolean gestureIsRight = posY > height * (1 - posX / width);
-        if (posX > posY / height * width) {
-            if (gestureIsRight) {
-                return mGestureTapRight;
-            } else {
-                return mGestureTapTop;
-            }
-        } else {
-            if (gestureIsRight) {
-                return mGestureTapBottom;
-            } else {
-                return mGestureTapLeft;
-            }
-        }
-    }
-
 
     @ViewerCommand.ViewerCommandDef
     public int getDoubleTap() {
@@ -133,85 +53,5 @@ public class GestureProcessor {
     @ViewerCommand.ViewerCommandDef
     public int getCommandFromFling(float dx, float dy, float velocityX, float velocityY, boolean isSelecting, boolean isXScrolling, boolean isYScrolling) {
         return this.mGestureMapper.gesture(dx, dy, velocityX, velocityY, isSelecting, isXScrolling, isYScrolling);
-    }
-
-
-    public enum GestureSegment {
-        TOP_LEFT,
-        TOP_CENTER,
-        TOP_RIGHT,
-        MIDDLE_LEFT,
-        MIDDLE_CENTER,
-        MIDDLE_RIGHT,
-        BOTTOM_LEFT,
-        BOTTOM_CENTER,
-        BOTTOM_RIGHT;
-
-        public static GestureSegment fromTap(int height, int width, float posX, float posY) {
-
-            double heightSegment = height / 3d;
-            double widthSegment = width / 3d;
-
-            TriState wSector = clamp(posX / widthSegment);
-            TriState hSector = clamp(posY / heightSegment);
-
-            switch (wSector) {
-                case LOW:
-                    //left
-                    switch (hSector) {
-                        case LOW:
-                            return TOP_LEFT;
-                        case MID:
-                            return MIDDLE_LEFT;
-                        case HIGH:
-                            return BOTTOM_LEFT;
-                    }
-                    break;
-                case MID:
-                    //center
-                    switch (hSector) {
-                        case LOW:
-                            return TOP_CENTER;
-                        case MID:
-                            return MIDDLE_CENTER;
-                        case HIGH:
-                            return BOTTOM_CENTER;
-                    }
-                    break;
-                case HIGH:
-                    //Right
-                    switch (hSector) {
-                        case LOW:
-                            return TOP_RIGHT;
-                        case MID:
-                            return MIDDLE_RIGHT;
-                        case HIGH:
-                            return BOTTOM_RIGHT;
-                    }
-                    break;
-                default:
-                    throw new IllegalArgumentException("illegal switch state");
-            }
-            throw new IllegalArgumentException("illegal switch state");
-        }
-
-        // clamps the value from LOW-MID-HIGH
-        private static TriState clamp(double value) {
-            double val = Math.floor(value);
-            if (val >= 2) {
-                return TriState.HIGH;
-            }
-            if (val < 1) {
-                return TriState.LOW;
-            }
-
-            return TriState.MID;
-        }
-
-        private enum TriState {
-            LOW,
-            MID,
-            HIGH
-        }
     }
 }
